cellViewModel refector 완료
