+++ VocaWeaveApp/Presenter/DictionaryViewPage/DictionaryVC.swift
@@ -59,12 +59,12 @@ final class DictionaryVC: UIViewController {
         if #available(iOS 13.0, *) {
             if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene {
                 if let window = windowScene.windows.first {
-                    if window.overrideUserInterfaceStyle == .dark {
-                        button.image = UIImage(systemName: "moon.fill")
-                        button.tintColor = .subTinkColor
-                    } else {
+                    if window.overrideUserInterfaceStyle == .light {
                         button.image = UIImage(systemName: "moon")
-                        button.tintColor = .black
+                        button.tintColor = .label
+                    } else {
+                        button.image = UIImage(systemName: "moon.fill")
+                        button.tintColor = UIColor(red: 0.91, green: 0.77, blue: 0.42, alpha: 1.00)
                     }
                 }
             }
